final_HW_module3/requirements.txt
apache-airflow==2.7.1
pymongo==4.5.0
psycopg2-binary==2.9.9
pandas==2.0.3
apache-airflow-providers-mongo==3.3.0
apache-airflow-providers-postgres==5.7.0
faker==20.1.0